Add new health endpoint with checkers
Refactored routes to pass in configuration variables than the handlers calling off to config.Get. Which would pick up changes to the environmental configuration and throw an error which was handled by calling os.Exit(). This would cause the application to kill itself uneccessarily.

Pulled in a dependency fix for converting metadata to string which was causing tests to fail due to not handling nil pointers and resulting in a panic.