Use 'printf' to add build flags when generating assets
Running zsh on a linux machine the `\n` (newline) character was being
added to the `data.go` file when using the `echo` command.
Additionally the `echo -e` of `echo -E` options for backslash
interpretation are not available on Darwin/bsd.

This is useful when running the commands:
```
make generate-debug
make generate-prod
```